requirements.txt
fastapi
uvicorn
python-multipart
demucs
torchaudio
pywebview
